refactor(CodeEditor): update import paths for testing utilities
- Changed import paths in various extension spec files to reflect the new location of testing utilities, improving organization and accessibility.
- Ensured consistency across all extension tests by standardizing the import structure.
