+++ packages/code-editor/src/CodeEditor/hooks/extensions/useAutoCompleteExtension.spec.ts
@@ -3,7 +3,7 @@ import { renderHook } from '@testing-library/react';
 import {
   createMockAutoCompleteModule,
   createMockStateModule,
-} from '../../testing';
+} from '../../../testing';
 
 import { useAutoCompleteExtension } from './useAutoCompleteExtension';
