+++ packages/code-editor/src/CodeEditor/hooks/extensions/useLanguageExtension.spec.ts
@@ -1,6 +1,6 @@
 import { renderHook } from '@testing-library/react';
 
-import { createMockStateModule } from '../../testing';
+import { createMockStateModule } from '../../../testing';
 
 import { LanguageName, useLanguageExtension } from './useLanguageExtension';
