+++ packages/code-editor/src/CodeEditor/hooks/extensions/useReadOnlyExtension.spec.ts
@@ -1,6 +1,6 @@
 import { renderHook } from '@testing-library/react';
 
-import { createMockStateModule } from '../../testing';
+import { createMockStateModule } from '../../../testing';
 
 import { useReadOnlyExtension } from './useReadOnlyExtension';
